test1

+++ Services/Repositories/RepoProgram.cs
@@ -1,6 +1,6 @@
 ﻿namespace liaqati_master.Services.RepoCrud
 {
-    public interface RepoProgram
+    public interface IRepoProgram
     {
         public Task<bool> AddProgram(SportsProgram SportsProgram);
         public Task<bool> UpdateProgram(SportsProgram SportsProgram);
@@ -12,38 +12,26 @@
 
 
 
-    public class ProgramMang : RepoProgram
+    public class ProgramMang : IRepoProgram
     {
-        public readonly LiaqatiDBContext _context;
+        private readonly LiaqatiDBContext _context;
 
         public ProgramMang(LiaqatiDBContext context)
         {
             _context = context;
         }
-
-
-
-
-
-
-
-
         public async Task<bool> AddProgram(SportsProgram SportsProgram)
         {
             await _context.TblSportsProgram.AddAsync(SportsProgram);
             await _context.SaveChangesAsync();
             return true;
         }
-
-
-
         public async Task<bool> DeleteProgram(string? id)
         {
             List<SportsProgram> SportsProgram = _context.TblSportsProgram.Include(x => x.Services).Include(p => p.Exercies_Programs).ToList();
-            SportsProgram sport = SportsProgram.Where(p => p.Equals(id)).ToList()[0];
-
+            SportsProgram? sport = SportsProgram.FirstOrDefault(p => p.Equals(id));
+            if (sport == null) { return false; }
             _context.TblSportsProgram.Remove(sport);
-
             await _context.SaveChangesAsync();
 
             return true;
@@ -57,23 +45,16 @@
 
             return SportsProgram;
         }
-
-
-
         public async Task<SportsProgram> GetProgram(string? id)
         {
             List<SportsProgram> SportsProgram = _context.TblSportsProgram.Include(x => x.Services).Include(p => p.Exercies_Programs).ToList();
 
-            SportsProgram sport = SportsProgram.Where(p => p.Id.Equals(id)).ToList()[0];
+            SportsProgram? sport = SportsProgram.FirstOrDefault(p => p.Id.Equals(id));
 
             await _context.SaveChangesAsync();
 
-
             return sport;
         }
-
-
-
 
         public async Task<bool> UpdateProgram(SportsProgram SportsProgram)
         {
